requirements.txt
discord.py==2.3.2
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
cchardet==2.1.7
gunicorn==21.2.0